Additional changes

+++ main.ts
@@ -11,22 +11,32 @@ input.onButtonPressed(Button.A, function () {
     GameStarted = 1
 })
 input.onButtonPressed(Button.B, function () {
-    Shoot = game.createSprite(Player.get(LedSpriteProperty.X), Player.get(LedSpriteProperty.Y))
-    Shoot.change(LedSpriteProperty.Brightness, 255)
-    for (let index = 0; index < 4; index++) {
-        Shoot.change(LedSpriteProperty.Y, -1)
-        basic.pause(100)
+    if (GameStarted == 1) {
+        Shoot = game.createSprite(Player.get(LedSpriteProperty.X), Player.get(LedSpriteProperty.Y))
+        Shoot.change(LedSpriteProperty.Brightness, 255)
+        for (let index = 0; index < 4; index++) {
+            Shoot.change(LedSpriteProperty.Y, -1)
+            if (Shoot.isTouching(Enemy)) {
+                Enemies = 0
+                game.addScore(1)
+                Enemy.delete()
+                Shoot.delete()
+            }
+            basic.pause(100)
+        }
+        Shoot.delete()
     }
-    Shoot.delete()
 })
 let Enemy: game.LedSprite = null
 let Shoot: game.LedSprite = null
 let Player: game.LedSprite = null
+let Enemies = 0
 let GameStarted = 0
 let dirPlayer = 0
 dirPlayer = 1
 GameStarted = 0
-let Enemies = 0
+Enemies = 0
+game.setScore(0)
 basic.forever(function () {
     if (GameStarted == 1) {
         Player.move(dirPlayer)
